+++ packages/building-webpack/demo/ts/webpack.config.js
@@ -1,10 +1,11 @@
+// @ts-nocheck
+
 const path = require('path');
 const merge = require('webpack-merge');
 const createDefaultConfig = require('../../modern-and-legacy-config');
 
 const defaultConfig = createDefaultConfig({
-  entry: path.resolve(__dirname, './demo-app.ts'),
-  indexHTML: path.resolve(__dirname, './index.html'),
+  input: path.resolve(__dirname, './index.html'),
 });
 
 module.exports = defaultConfig.map(config =>