+++ packages/create/src/generators/building-webpack/templates/static/webpack.config.js
@@ -1,3 +1,9 @@
-const defaultConfig = require('@open-wc/building-webpack/default-config');
+const { resolve } = require('path');
+const createDefaultConfig = require('@open-wc/building-webpack/modern-and-legacy-config');
 
-module.exports = defaultConfig();
+// If you don't need IE11 support, use the modern-config instead
+// import createDefaultConfig from '@open-wc/building-webpack/modern-config';
+
+module.exports = createDefaultConfig({
+  input: resolve(__dirname, './src/index.html'),
+});