+++ packages/semantic-dom-diff/src/get-dom-diff.d.ts
@@ -1,4 +1,0 @@
-import { DocumentFragment } from 'parse5';
-
-export function getAST(value, config): DocumentFragment
-export function getDiffableSemanticHTML(html: string): string