+++ packages/testing-helpers/lit-html.d.ts
@@ -1,2 +1,0 @@
-export { render, html } from 'lit-html/lit-html';
-export function unsafeStatic<TOptions>(options: TOptions): { d: TOptions }