+++ packages/testing-helpers/litFixture.d.ts
@@ -1,4 +1,0 @@
-import { TemplateResult } from 'lit-html';
-
-export function litFixtureSync<T extends Element = Element>(template: TemplateResult): T;
-export function litFixture<T extends Element = Element>(template: TemplateResult): Promise<T>;