+++ packages/testing-helpers/test/lit-html.test.js
@@ -1,12 +1,19 @@
 import { expect } from '@bundled-es-modules/chai';
-import { html, litFixture, unsafeStatic } from '../index.js';
+import { html, litFixture, unsafeStatic } from '..';
+
+/**
+ * @typedef {Object} ChildType
+ * @property {number} myNumber
+ *
+ * @typedef {Element & ChildType} TestDiv
+ */
 
 describe('html', () => {
   it('renders dynamic tags', async () => {
     const tag = unsafeStatic('my-foo');
-    const el = await litFixture(html`
+    const el = /** @type {TestDiv} */ (await litFixture(html`
       <${tag} .myNumber=${4} foo="bar"></${tag}>
-    `);
+    `));
     expect(el.tagName).to.equal('MY-FOO');
     expect(el.myNumber).to.equal(4);
     expect(el.getAttribute('foo')).to.equal('bar');
@@ -20,9 +27,9 @@ describe('html', () => {
   });
 
   it('renders static templates with properties, attributes', async () => {
-    const el = await litFixture(html`
+    const el = /** @type {TestDiv} */ (await litFixture(html`
       <div .myNumber=${4} foo="bar"></div>
-    `);
+    `));
     expect(el.tagName).to.equal('DIV');
     expect(el.myNumber).to.equal(4);
     expect(el.getAttribute('foo')).to.equal('bar');