+++ packages/building-rollup/CHANGELOG.md
@@ -3,6 +3,104 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.2.4...@open-wc/building-rollup@0.3.0) (2019-04-08)
+
+
+### Features
+
+* **semantic-dom-diff:** add support for snapshot testing ([f7a675a](https://github.com/open-wc/open-wc/commit/f7a675a))
+
+
+
+
+
+## [0.2.4](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.2.3...@open-wc/building-rollup@0.2.4) (2019-04-05)
+
+
+### Bug Fixes
+
+* do not assume available global types of users ([cd394d9](https://github.com/open-wc/open-wc/commit/cd394d9))
+
+
+
+
+
+## [0.2.3](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.2.2...@open-wc/building-rollup@0.2.3) (2019-04-03)
+
+
+### Bug Fixes
+
+* add clone dependency ([4d15a39](https://github.com/open-wc/open-wc/commit/4d15a39))
+
+
+
+
+
+## [0.2.2](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.2.1...@open-wc/building-rollup@0.2.2) (2019-03-27)
+
+
+### Bug Fixes
+
+* **building-rollup:** Add missing dependency ([#326](https://github.com/open-wc/open-wc/issues/326)) ([60d0fec](https://github.com/open-wc/open-wc/commit/60d0fec))
+
+
+
+
+
+## [0.2.1](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.2.0...@open-wc/building-rollup@0.2.1) (2019-03-23)
+
+
+### Bug Fixes
+
+* **building-rollup:** include dom5 fork in npm bundle ([0909aa6](https://github.com/open-wc/open-wc/commit/0909aa6))
+
+
+
+
+
+# [0.2.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.1.10...@open-wc/building-rollup@0.2.0) (2019-03-23)
+
+
+### Bug Fixes
+
+* **building-rollup:** temporary fork dom5 as it has outdated typings ([3d0d187](https://github.com/open-wc/open-wc/commit/3d0d187))
+
+
+### Features
+
+* add types + linting & improve intellisense ([b6d260c](https://github.com/open-wc/open-wc/commit/b6d260c))
+
+
+
+
+
+## [0.1.10](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.1.9...@open-wc/building-rollup@0.1.10) (2019-03-23)
+
+
+### Bug Fixes
+
+* **building-rollup:** disable http server caching ([#313](https://github.com/open-wc/open-wc/issues/313)) ([d3ce2b2](https://github.com/open-wc/open-wc/commit/d3ce2b2))
+
+
+
+
+
+## [0.1.9](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.1.8...@open-wc/building-rollup@0.1.9) (2019-03-23)
+
+**Note:** Version bump only for package @open-wc/building-rollup
+
+
+
+
+
+## [0.1.8](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.1.7...@open-wc/building-rollup@0.1.8) (2019-03-23)
+
+**Note:** Version bump only for package @open-wc/building-rollup
+
+
+
+
+
 ## [0.1.7](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.1.6...@open-wc/building-rollup@0.1.7) (2019-03-22)