+++ packages/polyfills-loader/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/open-wc/open-wc/compare/@open-wc/polyfills-loader@0.3.0...@open-wc/polyfills-loader@0.3.1) (2019-04-08)
+
+
+### Bug Fixes
+
+* **polyfills-loader:** typo in readme ([604c4de](https://github.com/open-wc/open-wc/commit/604c4de))
+
+
+
+
+
+# [0.3.0](https://github.com/open-wc/open-wc/compare/@open-wc/polyfills-loader@0.2.2...@open-wc/polyfills-loader@0.3.0) (2019-03-23)
+
+
+### Features
+
+* add types + linting & improve intellisense ([b6d260c](https://github.com/open-wc/open-wc/commit/b6d260c))
+
+
+
+
+
 ## [0.2.2](https://github.com/open-wc/open-wc/compare/@open-wc/polyfills-loader@0.2.1...@open-wc/polyfills-loader@0.2.2) (2019-02-16)