+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,158 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.2](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.11.1...@open-wc/testing@0.11.2) (2019-04-08)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.11.1](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.11.0...@open-wc/testing@0.11.1) (2019-04-08)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+# [0.11.0](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.12...@open-wc/testing@0.11.0) (2019-04-08)
+
+
+### Features
+
+* **semantic-dom-diff:** add support for snapshot testing ([f7a675a](https://github.com/open-wc/open-wc/commit/f7a675a))
+
+
+
+
+
+## [0.10.12](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.11...@open-wc/testing@0.10.12) (2019-04-06)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.10.11](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.10...@open-wc/testing@0.10.11) (2019-04-05)
+
+
+### Bug Fixes
+
+* do not assume available global types of users ([cd394d9](https://github.com/open-wc/open-wc/commit/cd394d9))
+
+
+
+
+
+## [0.10.10](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.9...@open-wc/testing@0.10.10) (2019-04-03)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.10.9](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.8...@open-wc/testing@0.10.9) (2019-03-31)
+
+
+### Bug Fixes
+
+* **semantic-dom-diff:** add get-diffable-html.js to npm bundle ([7ee3ba9](https://github.com/open-wc/open-wc/commit/7ee3ba9))
+
+
+
+
+
+## [0.10.8](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.7...@open-wc/testing@0.10.8) (2019-03-31)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.10.7](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.6...@open-wc/testing@0.10.7) (2019-03-31)
+
+
+### Bug Fixes
+
+* **testing:** update instructions for new karma config setup ([2de10bf](https://github.com/open-wc/open-wc/commit/2de10bf))
+
+
+
+
+
+## [0.10.6](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.5...@open-wc/testing@0.10.6) (2019-03-31)
+
+
+### Bug Fixes
+
+* adopt new karma setup for all packages ([1888260](https://github.com/open-wc/open-wc/commit/1888260))
+
+
+
+
+
+## [0.10.5](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.4...@open-wc/testing@0.10.5) (2019-03-29)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.10.4](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.3...@open-wc/testing@0.10.4) (2019-03-28)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.10.3](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.2...@open-wc/testing@0.10.3) (2019-03-27)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.10.2](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.1...@open-wc/testing@0.10.2) (2019-03-24)
+
+
+### Bug Fixes
+
+* adjust generator-open-wc links to create ([cc014b1](https://github.com/open-wc/open-wc/commit/cc014b1))
+
+
+
+
+
+## [0.10.1](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.10.0...@open-wc/testing@0.10.1) (2019-03-23)
+
+
+### Bug Fixes
+
+* do not assume globally setup mocha types ([977d5b4](https://github.com/open-wc/open-wc/commit/977d5b4))
+
+
+
+
+
+# [0.10.0](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.9.28...@open-wc/testing@0.10.0) (2019-03-23)
+
+
+### Features
+
+* add types + linting & improve intellisense ([b6d260c](https://github.com/open-wc/open-wc/commit/b6d260c))
+
+
+
+
+
 ## [0.9.28](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.9.27...@open-wc/testing@0.9.28) (2019-03-20)
 
 **Note:** Version bump only for package @open-wc/testing